Separating styles
Created a new file for the styles of outdoor directions components since multiple files share similar styles
